Fix getAllSelfTickets code ambiguity
The current code in the guide is misleading by having a Pageable (unused),and a paramater "eagerload", which breaks getAllSelfTickets contract. If "eagerload" is set to true, the method won't respect its contract, which is to give the user's assigned tickets, and will instead give all existing tickets.

Since I can't find a simple clear way to make the user's assigned tickets pageable (without modifying TicketRepository), I just removed most of the code, but kept the Pagination header, to keep the same response format as before.